more lint fixes, todo update logics fix
