--- .github/workflows/build-tmp.yml
@@ -1,74 +1,0 @@
-name: build
-
-on:
-  push:
-    branches:
-      - erhant/rewrite-in-rust
-
-jobs:
-  ########################################
-  # Builds the releases for all platforms
-  ########################################
-  build:
-    runs-on: ${{ matrix.runner }}
-    strategy:
-      matrix:
-        include:
-          # macOS intel
-          - {
-              runner: macos-latest,
-              osname: macOS,
-              arch: amd64,
-              target: x86_64-apple-darwin,
-            }
-          # macOS arm (apple silicon)
-          - {
-              runner: macos-latest,
-              osname: macOS,
-              arch: arm64,
-              target: aarch64-apple-darwin,
-            }
-          # linux amd64
-          - {
-              runner: ubuntu-latest,
-              osname: linux,
-              arch: amd64,
-              target: x86_64-unknown-linux-musl,
-            }
-          # linux arm64
-          - {
-              runner: ubuntu-latest,
-              osname: linux,
-              arch: arm64,
-              target: aarch64-unknown-linux-musl,
-            }
-          # windows x86_64
-          - {
-              runner: windows-latest,
-              osname: windows,
-              arch: amd64,
-              target: x86_64-pc-windows-msvc,
-              extension: ".exe",
-            }
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      - name: Get tag
-        shell: bash
-        run: echo "VERSION=${GITHUB_REF#refs/tags/}" >> $GITHUB_ENV
-
-      - name: Build binary
-        uses: houseabsolute/actions-rust-cross@v0
-        with:
-          command: "build"
-          target: ${{ matrix.target }}
-          args: "--locked --release"
-          strip: true
-
-      - name: Upload Launch Artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          name: dkn-compute-launcher-${{ matrix.osname }}-${{ matrix.arch }}${{ matrix.extension }}
-          path: target/${{ matrix.target }}/release/dkn-compute-launcher${{ matrix.extension }}
