+++ .github/workflows/tests.yml
@@ -6,6 +6,7 @@ on:
       - master
     paths:
       - "src/**"
+      - "tests/**"
       - "Cargo.lock"
       - ".github/workflows/tests.yml"
   workflow_dispatch:
