+++ src/utils/mod.rs
@@ -9,3 +9,6 @@ pub use env::*;
 
 mod process;
 pub use process::*;
+
+mod updates;
+pub use updates::*;
